run tests against http://httpbin.org
this makes the tests run against http://httpbin.org instead of my homegrown
sinatra app. This is better for a couple of reasons:

- httpbin supports more features so it can be used for future development as
  well
- it actually works and I don't have to maintain it
- Python's wonderful requests library also uses it so we know it's good

There is however a downside to it. As the test suite right now doesn't support
JSON parsing, I had to remove some unit tests until I add that. I think that's
still fine, but I opened issue #23 to address this.

This fixes #12