add vendorized json-cpp
this is a checking of the [amalgamated][1] source of the json-cpp parser. It's
a great parser and this way we can just include it in the test build without
having to build another vendor library and dealing with their build system.
This gives us the ability to parse output from httpbin for better unit tests.

[1]: https://github.com/open-source-parsers/jsoncpp#generating-amalgamated-source-and-header