remove calls to curl_global_cleanup
this method is not threadsafe and should not be called from a library. It
cleans up a lot of non-local setups and thus should be called from the
application. Maybe it's worth adding a cleanup function to restclient for this
to make it more explicit that a global cleanup needs to be done by the caller.
However it would only call curl_global_cleanup right now anyways.

reference: http://curl.haxx.se/mail/lib-2012-05/0034.html
fixes #19 #29