cleanup version gen to be driven by git tags
this removes the static meta.h header file which included only version
information and replaces it with a version.h file. This file is generated on
build time from the m4/version.h.m4 template and includes the version string
as generated from `git describe --tags --always --dirty`. This makes sure that
the version of any builds always corresponds to an existing revision in git
(and also marks it as dirty if the build comes from uncommited changes).