Added a method to perform a post form upload.
Fixed code indentation to comply with Cpplint.

Added web proxy tunneling support.

this adds support for using proxies corresponding to the CURLOPT_PROXY
and CURLOPT_HTTPPROXYTUNNEL config options in libcurl.

closes #68

default to 10s timeout in unit tests

this makes the conn object in unit tests default to a 10s timeout in
order to fail fast if something is up with the connection or configured
hosts.