mac os x uses glibtoolize not libtoolize
for mac os x hosts "libtoolize" is not a correct command.
this is because mac bundles it's own version of "libtool",
which works differently, and may not even exist.

instead if you use brew to install libtool: `brew install libtool`
on mac, it writes to: `glibtoolize` as to leave mac happy with
it's own version of libtool by default.

this commit will properly use glibtoolize if detected, otherwise
will fallback to just using libtoolize.