Don't half-handle cURL error codes (closes #153)
If cURL returns some error code, just return that to the caller as
ret.code. This works because CURLE codes currently stop at 92 (i.e. they
do not overlap with HTTP 1xx codes).

If a caller notices a code 92 or below, they can inspect ret.body for
some rough analysis of the error. If that is not good enough, they can
also inspect GetInfo().lastRequest.curlError to see the output that curl
put in the curl error buffer (which is for some reason different than
curl_easy_strerror output).